+++ CHANGELOG.md
@@ -1,13 +1,37 @@
 # Sourcery CHANGELOG
 
 ---
-## Master
+## 0.5.2
+
+### New Features
+
+- Added support for `ImplicitlyUnwrappedOptional`
+- `actualTypeName` property of `Method.Parameter`, `Variable`, `Enum.Case.AssociatedValue`, `TupleType.Element` now returns `typeName` if type is not a type alias
+- `Enum` now contains type information for its raw value type. `rawType` now return `Type` object, `rawTypeName` returns its `TypeName`
+- Added `annotated` filter to filter by annotations
+- Added negative filters counterparts
+- Added support for attributes, i.e. `@escaping`
+- Experimental support for Swift Templates
+
+```
+<% for type in types.classes { %>
+    extension <%= type.name %>: Equatable {}
+
+    <% if type.annotations["showComment"] != nil { %> // <%= type.name %> has Annotations <% } %>
+
+        func == (lhs: <%= type.name %>, rhs: <%= type.name %>) -> Bool {
+    <% for variable in type.variables { %> if lhs.<%= variable.name %> != rhs.<%= variable.name %> { return false }
+        <% } %>
+        return true
+    }
+<% } %>
+```
 
 ### New Features
 - Variables with default initializer are now supported, e.g. `var variable = Type(...)`
 - Added support for special escaped names in enum cases e.g. `default` or `for`
 - Added support for tuple types and `tuple` filter for variables
-- Enum associated values now have `localName` and `exteranlName` properties.
+- Enum associated values now have `localName` and `externalName` properties.
 - Added `actualTypeName` for `TypeName` that is typealias
 - Added `implements`, `inherits` and `based` filters